Group javax.jms instrumentations under a single module
