+++ dd-java-agent/instrumentation/jms/src/main/java/datadog/trace/instrumentation/jms/SessionInstrumentation.java
@@ -13,17 +13,13 @@ import static net.bytebuddy.matcher.ElementMatchers.isPublic;
 import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
 import static net.bytebuddy.matcher.ElementMatchers.takesNoArguments;
 
-import com.google.auto.service.AutoService;
 import datadog.trace.agent.tooling.Instrumenter;
-import datadog.trace.agent.tooling.InstrumenterModule;
 import datadog.trace.api.Config;
 import datadog.trace.bootstrap.ContextStore;
 import datadog.trace.bootstrap.InstrumentationContext;
 import datadog.trace.bootstrap.instrumentation.jms.MessageConsumerState;
 import datadog.trace.bootstrap.instrumentation.jms.MessageProducerState;
 import datadog.trace.bootstrap.instrumentation.jms.SessionState;
-import java.util.HashMap;
-import java.util.Map;
 import javax.jms.Destination;
 import javax.jms.MessageConsumer;
 import javax.jms.MessageProducer;
@@ -32,12 +28,8 @@ import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
 
-@AutoService(InstrumenterModule.class)
-public class SessionInstrumentation extends InstrumenterModule.Tracing
+public class SessionInstrumentation
     implements Instrumenter.ForTypeHierarchy, Instrumenter.HasMethodAdvice {
-  public SessionInstrumentation() {
-    super("jms", "jms-1", "jms-2");
-  }
 
   @Override
   public String hierarchyMarkerType() {
@@ -47,20 +39,6 @@ public class SessionInstrumentation extends InstrumenterModule.Tracing
   @Override
   public ElementMatcher<TypeDescription> hierarchyMatcher() {
     return implementsInterface(named(hierarchyMarkerType()));
-  }
-
-  @Override
-  public String[] helperClassNames() {
-    return new String[] {packageName + ".JMSDecorator"};
-  }
-
-  @Override
-  public Map<String, String> contextStore() {
-    Map<String, String> contextStore = new HashMap<>(4);
-    contextStore.put("javax.jms.MessageConsumer", MessageConsumerState.class.getName());
-    contextStore.put("javax.jms.MessageProducer", MessageProducerState.class.getName());
-    contextStore.put("javax.jms.Session", SessionState.class.getName());
-    return contextStore;
   }
 
   @Override
